refactor: Organização do Controller na camada de apresentação
